BLE: decouple the BLE listening mode handler.
